fix merge  error
use of new _instantiate_new_instance_from_db() for obtaining objects was  overwritten by old  code that used  EE_Registry. This obviously  happened during a merge but does not  appear in the log. I'm assuming the conflict was because of incoming PSR formatting and someone didn't look over the changes carfeully and then resolved in favour of master, nor did they touch the ticket. [touch:9856]